+++ LiveCD/livecd.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ "$1" = "CI" ]; then
+    OPT='-dR'
+else
+    OPT='-a'
+fi
+
 VENTOY_PATH=$PWD/../
 
 if ! [ -f $VENTOY_PATH/INSTALL/grub/grub.cfg ]; then
@@ -15,18 +21,33 @@ if ! [ -e $VENTOY_PATH/INSTALL/ventoy-${version}-linux.tar.gz ]; then
 fi
 
 rm -rf ISO_TMP
-cp -a ISO ISO_TMP
+cp $OPT ISO ISO_TMP
 
-cp -a VTOY VTOY_TMP && cd VTOY_TMP
+cp $OPT VTOY VTOY_TMP 
+
+ls -la
+if ! [ -d ISO_TMP ]; then
+    echo "Copy ISO_TMP failed"
+    exit 1
+fi
+
+if ! [ -d VTOY_TMP ]; then
+    echo "Copy VTOY_TMP failed"
+    exit 1
+fi
+
+mkdir -p ISO_TMP/EFI/ventoy
+cd VTOY_TMP
+
 gcc -O2 -m32 ./ventoy/disksize.c -o ./ventoy/disksize
 rm -f ./ventoy/disksize.c
-find . | cpio  -o -H newc | gzip -9 > ../ISO_TMP/EFI/ventoy/ventoy.gz
+find . | cpio  -o -H newc | gzip -c -9 > ../ISO_TMP/EFI/ventoy/ventoy.gz
 cd .. && rm -rf VTOY_TMP
 
 
-cp -a $VENTOY_PATH/INSTALL/ventoy-${version}-linux.tar.gz ISO_TMP/EFI/ventoy/
-cp -a GRUB/cdrom.img ISO_TMP/EFI/boot/
-cp -a GRUB/bootx64.efi ISO_TMP/EFI/boot/
+cp $OPT $VENTOY_PATH/INSTALL/ventoy-${version}-linux.tar.gz ISO_TMP/EFI/ventoy/
+cp $OPT GRUB/cdrom.img ISO_TMP/EFI/boot/
+cp $OPT GRUB/bootx64.efi ISO_TMP/EFI/boot/
 
 
 rm -rf efimnt
@@ -37,11 +58,11 @@ dd if=/dev/zero of=efi.img bs=1M count=2
 mkfs.vfat efi.img
 mount efi.img efimnt
 mkdir -p efimnt/EFI/boot
-cp -a GRUB/bootx64.efi efimnt/EFI/boot/
+cp $OPT GRUB/bootx64.efi efimnt/EFI/boot/
 umount efimnt
 
 sync
-cp -a efi.img ISO_TMP/EFI/boot/
+cp $OPT efi.img ISO_TMP/EFI/boot/
 
 rm -rf efimnt
 rm -f efi.img
